nessus_api/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
celery>=5.0,<6.0
redis>=4.0,<5.0
django-cors-headers>=3.10,<4.0
